Separated db.py into emoji and title suggestion part
