+++ main.py
@@ -17,6 +17,7 @@ import temproles
 import threadwatch
 import timeywimey
 import yud
+import tbdtools
 from config import (
     BLUE_PORTAL,
     CATHEARTS,
@@ -36,6 +37,7 @@ from config import (
     LOAD_TEMPROLES,
     LOAD_THREADWATCH,
     LOAD_YUD,
+    LOAD_TBDTOOLS,
     LOGGER_CHANNEL,
     ORANGE_PORTAL,
     PLEADING,
@@ -78,9 +80,12 @@ async def on_ready():
     # !remindme
     if LOAD_REMINDERS:
         await TBotD.add_cog(reminders.Reminders(TBotD, connection))
-    # Store "TBD" title suggestions, and used emoji status (for no real reason)
+    # Counts the number of reacted emojis
     if LOAD_DB:
         await TBotD.add_cog(db.Database(TBotD, connection))
+    # Keeps track of TBD title suggestions
+    if LOAD_TBDTOOLS:
+        await TBotD.add_cog(tbdtools.TBDTools(TBotD, connection))
     # !cwbanme and related commands
     if LOAD_TEMPROLES:
         await TBotD.add_cog(temproles.RoleManagement(TBotD, tbd, connection))
